remove mergeObjects codepaths and rename file
